requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
channels==4.1.0
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.1.2
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-rest-knox==5.0.2
djangorestframework==3.16.0
google-auth==2.34.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
openai==1.107.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyOpenSSL==25.3.0
requests==2.32.3
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zope.interface==8.0.1
